requirements.txt
# Core Dependencies
scikit-learn>=1.0
tensorflow>=2.8
pandas>=1.3
numpy>=1.21
scipy>=1.7

# Network & Data Processing
scapy>=2.4
pyshark>=0.4
kafka-python>=2.0
pyspark>=3.2
elasticsearch>=7.0
pymongo>=4.0
redis>=4.0

# Web & API
fastapi>=0.68
uvicorn>=0.15
requests>=2.26
aiohttp>=3.8
websockets>=10.0

# Utilities
python-dotenv>=0.19
pyyaml>=6.0
loguru>=0.6
tqdm>=4.62
joblib>=1.1

# Security & Cryptography
cryptography>=36.0
pyjwt>=2.3
bcrypt>=3.2
python-multipart>=0.0

# ML & Data Science
xgboost>=1.5
lightgbm>=3.3
catboost>=1.0
imbalanced-learn>=0.9
shap>=0.40

# Threat Intelligence
pyattck>=6.0
# otx-sdk>=1.5
# abuseipdb>=0.1

# Testing & Development
pytest>=6.2
pytest-asyncio>=0.16
pytest-cov>=3.0
black>=22.0
flake8>=4.0
mypy>=0.910
